split tasks and made changes
